+++ bootstrap/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.5.5]
+
+- Set autonat to enabled, flag external_address parameter as mandatory
+- Update `avail-light-core` to 1.2.11
+
 ## [0.5.4](https://github.com/availproject/avail-light/releases/tag/avail-light-bootstrap-v0.5.4) - 2025-09-08
 
 - Update `avail-light-core` to 1.2.10